+++ src/commands/newConnection/CosmosDBExecuteStep.ts
@@ -12,7 +12,7 @@ import { WorkspaceResourceType } from '../../tree/workspace-api/SharedWorkspaceR
 import {
     SharedWorkspaceStorage,
     type SharedWorkspaceStorageItem,
-} from '../../tree/workspace-api/SharedWorkspaceStorage';
+} from '../../services/SharedWorkspaceStorage';
 import { type NewConnectionWizardContext } from './NewConnectionWizardContext';
 
 export class CosmosDBExecuteStep extends AzureWizardExecuteStep<NewConnectionWizardContext> {
