+++ src/commands/newConnection/MongoExecuteStep.ts
@@ -12,7 +12,7 @@ import { WorkspaceResourceType } from '../../tree/workspace-api/SharedWorkspaceR
 import {
     SharedWorkspaceStorage,
     type SharedWorkspaceStorageItem,
-} from '../../tree/workspace-api/SharedWorkspaceStorage';
+} from '../../services/SharedWorkspaceStorage';
 import { type NewConnectionWizardContext } from './NewConnectionWizardContext';
 
 export class MongoExecuteStep extends AzureWizardExecuteStep<NewConnectionWizardContext> {
